requirements.txt
asgiref==3.9.1
boto3==1.40.20
botocore==1.40.20
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-storages==1.14.6
dotenv==0.9.9
filelock==3.19.1
fsspec==2025.7.0
gunicorn==23.0.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
